+++ modelcraft/jobs/sheetbend.py
@@ -16,11 +16,13 @@ class Sheetbend(Job):
         structure: gemmi.Structure,
         fsigf: DataItem,
         freer: DataItem = None,
+        regularise: bool = False,
     ):
         super().__init__("csheetbend")
         self.structure = structure
         self.fsigf = fsigf
         self.freer = freer
+        self.regularise = regularise
 
     def _setup(self) -> None:
         write_mmcif(self._path("xyzin.cif"), self.structure)
@@ -33,6 +35,10 @@ class Sheetbend(Job):
         self._args += ["-pdbout", "xyzout.cif"]
         self._args += ["-cycles", "12"]
         self._args += ["-resolution-by-cycle", "6,3"]
+        if self.regularise:
+            self._args += ["-postrefine-u-iso"]
+            self._args += ["-pseudo-regularize"]
+            self._args += ["-refine-regularize-cycles", "3"]
 
     def _result(self) -> SheetbendResult:
         self._check_files_exist("xyzout.cif")